flow - export util functions, re-org src and stories
