+++ src/index.js
@@ -1,3 +1,2 @@
 export { default as Visual } from './visual'
 export { default as Flow } from './flow'
-export { transform } from './utils'
